assignment/requirements.txt
openai
pandas
tqdm
requests
google-play-scraper-py
python-dotenv 